+++ modules/ank/README.md
@@ -186,7 +186,6 @@ Example:
 ````
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.data.*
 import arrow.core.*
 
 DataType(Option::class).tcMarkdownList()
