+++ modules/aql/arrow-query-language/build.gradle
@@ -10,7 +10,6 @@ dependencies {
     compileOnly project(':arrow-meta')
     testCompileOnly project(':arrow-meta')
     testRuntime("org.junit.vintage:junit-vintage-engine:$jUnitVintageVersion")
-    
     testCompile "io.kotlintest:kotlintest-runner-junit4:$kotlinTestVersion"
     testCompile project(':arrow-test')
 }
