+++ modules/core/arrow-test/src/main/kotlin/arrow/test/laws/PrismLaws.kt
@@ -10,7 +10,6 @@ import arrow.core.extensions.const.applicative.applicative
 import arrow.core.extensions.id.applicative.applicative
 import arrow.core.orElse
 import arrow.core.value
-import arrow.data.value
 import arrow.optics.Prism
 import arrow.optics.modify
 import arrow.typeclasses.Const
