+++ modules/docs/arrow-docs/docs/docs/arrow/core/either/README.md
@@ -361,7 +361,6 @@ Either.fx {
 
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.data.*
 import arrow.core.*
 
 DataType(Either::class).tcMarkdownList()
