+++ modules/docs/arrow-docs/docs/docs/arrow/data/writert/README.md
@@ -15,7 +15,7 @@ advanced
 
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.data.*
+import arrow.mtl.*
 import arrow.core.*
 
 DataType(WriterT::class).tcMarkdownList()
