+++ modules/effects/arrow-effects-extensions/src/main/kotlin/arrow/effects/extensions/writert.kt
@@ -4,11 +4,11 @@ import arrow.Kind
 import arrow.core.Either
 import arrow.core.Tuple2
 import arrow.core.compose
-import arrow.data.WriterT
-import arrow.data.WriterTOf
-import arrow.data.WriterTPartialOf
-import arrow.data.extensions.WriterTMonadThrow
-import arrow.data.value
+import arrow.mtl.WriterT
+import arrow.mtl.WriterTOf
+import arrow.mtl.WriterTPartialOf
+import arrow.mtl.extensions.WriterTMonadThrow
+import arrow.mtl.value
 import arrow.effects.Ref
 import arrow.effects.typeclasses.Async
 import arrow.effects.typeclasses.Bracket
